+++ NEWS.md
@@ -10,6 +10,8 @@ Version 1.0.0b0 - dev
 - CHANGE: removed deprecated features
 - BUGFIX: [#747](https://github.com/mozman/ezdxf/issues/747)
   fix virtual entities of 3D DIMENSION entities  
+- BUGFIX: [#748](https://github.com/mozman/ezdxf/issues/748)
+  fix keyword only argument in virtual_block_reference_entities() call
 
 Version 0.18.1 - 2022-09-03
 ---------------------------